feat: change js native to react
