+++ package.json
@@ -1,4 +1,5 @@
 {
+  "homepage": "https://iq-developer.github.io/react-game",
   "name": "react-game",
   "version": "0.1.0",
   "private": true,
@@ -13,10 +14,12 @@
   },
   "scripts": {
     "start": "react-scripts start",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
+ },
   "eslintConfig": {
     "extends": [
       "react-app",
@@ -34,5 +37,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "gh-pages": "^3.1.0"
   }
 }
