+++ src/Footer.js
@@ -1,0 +1,9 @@
+function Footer() {
+  return (
+    <div className="footer">
+      <p>Developed in 2021-2022 by <a href="https://github.com/iq-developer">Vladimir Sevastianov</a> at <a href="https://rs.school/js/"><img alt="RS School logo" src="https://rs.school/images/rs_school_js.svg" width="100" /></a></p>
+    </div>
+  );
+}
+
+export default Footer;
