+++ src/helpers/randomInteger.js
@@ -1,0 +1,6 @@
+const randomInteger = (min, max) => {
+  const rand = min - 0.5 + Math.random() * (max - min + 1);
+  return Math.round(rand);
+}
+
+export default randomInteger;
